core-device/src/backend/requirements.txt
awsiotsdk==1.7.1
flask-socketio==5.1.1
flask-cors==5.0.0
RPi.GPIO==0.7.1
spidev==3.6
Flask-SQLAlchemy==3.0.5
eventlet==0.34.1